added sorting by columns
