+++ HM12_algos2_buble_sort_etc/Sorting.h
@@ -1,0 +1,13 @@
+#pragma once
+
+constexpr unsigned int ROWS = 5;
+constexpr unsigned int COLUMNS = 6;
+
+enum class SortingDirection
+{
+    ByRows,
+    ByColumns
+};
+
+void bubbleSort(int arr[], int size);
+void sort(int arr[ROWS][COLUMNS], SortingDirection SortingDirection);
